plugins/agency-knows/skills/kb-github/scripts/pyproject.toml
[project]
name = "agency-knows"
version = "0.2.0"
description = "Generate and maintain knowledge base articles from source code"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "anthropic>=0.40",
    "httpx>=0.27",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "typer>=0.12",
]

[project.scripts]
agency-knows = "agency_knows.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agency_knows"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "C4", "SIM", "PLR"]
ignore = ["S603", "S607", "PLR0913", "PLR0912", "PLR0915", "PLR2004"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
